+++ examples/C++/DeadlineQoSExample/deadlinepayloadPublisher.h
@@ -46,7 +46,7 @@ public:
      * @param sleep_ms A time period to sleep for before sending the new sample
      * @param samples The number of samples per instance to send. If set to 0 sends sample indefinitely
      */
-    void run(double sleep_ms, int samples);
+    void run(uint32_t sleep_ms, int samples);
 
 private:
